Add files via upload
In this project about establishing the simulation system of forest runoff arrival time warning and prediction, there are three parts of code. The first part is c++ code written by esp32. nord has installed various sensors to measure various environments at the installation point, such as in him, in the middle and in the village below him. The second part is the website code, which shows the dashboard, tells the sensor values, reads the location of nord at different locations, and predicts forest runoff at the installation point and time of nord.